refactory code
