+++ backend/app/schemas/challenge/challenge_schema.py
