+++ backend/app/services/challenge/challenge_service.py
